Compatibility with TensorFlow v1.10.0

+++ api/src/main/scala/org/platanios/tensorflow/api/core/Shape.scala
@@ -345,7 +345,7 @@ final class Shape private (private val array: Array[Int]) extends ProtoSerializa
     * @param  dataType Data type to use for the tensor.
     * @return One-dimensional op output tensor representing this shape.
     */
-  def toOutput(dataType: DataType = INT32, name: String = "Shape"): Output = {
+  def toOutput(dataType: DataType = INT64, name: String = "Shape"): Output = {
     Basic.constant(toTensor(dataType), name = name)
   }
