+++ api/src/main/scala/org/platanios/tensorflow/api/tensors/Tensor.scala
@@ -288,7 +288,7 @@ class Tensor[+D <: DataType] protected (
     */
   override def toTensorIndexedSlices: TensorIndexedSlices[D] = {
     TensorIndexedSlices(
-      indices = Tensor(0L, 1L until shape(0).toLong: _*),
+      indices = Tensor(0, 1 until shape(0): _*),
       values = this,
       denseShape = shape.toTensor(INT64))
   }
@@ -705,9 +705,9 @@ final case class TensorIndexedSlices[+D <: DataType](
   * For example, the sparse tensor `SparseTensor(indices = [[0, 0], [1, 2]], values = [1, 2], denseShape = [3, 4])`,
   * represents the dense tensor `[[1, 0, 0, 0], [0, 0, 2, 0], [0, 0, 0, 0]]`.
   *
-  * @param  indices    Two-dimensional [[INT64]] tensor with shape `[N, rank]`.
+  * @param  indices    Two-dimensional [[INT32]] tensor with shape `[N, rank]`.
   * @param  values     One-dimensional tensor with shape `[N]`.
-  * @param  denseShape One-dimensional [[INT64]] tensor with shape `[rank]`.
+  * @param  denseShape One-dimensional [[INT32]] tensor with shape `[rank]`.
   *
   * @author Emmanouil Antonios Platanios
   */
