+++ api/src/main/scala/org/platanios/tensorflow/api/tensors/ops/NN.scala
@@ -464,7 +464,7 @@ private[api] trait NN {
     if (weights != null && weights.rank != 2)
       throw InvalidShapeException(s"'weights' must have shape [batchSize, sequenceLength], but had: ${weights.shape}.")
     val numClasses = Basic.shape(logits)(2)
-    val flattenedLogits = Basic.reshape(logits, Basic.stack(Seq[Tensor[INT32]](-1, numClasses)))
+    val flattenedLogits = Basic.reshape(logits, Basic.stack(Seq[Tensor[INT64]](-1L, numClasses)))
     val flattenedLabels = Basic.reshape(labels, Shape(-1))
     val epsilon = 1e-12.toTensor.cast(logits.dataType)
     var loss = lossFn(flattenedLogits, flattenedLabels)
@@ -624,7 +624,7 @@ private[api] trait NN {
 
   //region Convolution Ops
 
-  /** $OpDocConv2D
+  /** $OpDocNNConv2D
     *
     * @group NNOps
     * @param  input         4-D tensor whose dimension order is interpreted according to the value of `dataFormat`.
@@ -659,7 +659,7 @@ private[api] trait NN {
       Array(dilations._1, dilations._2, dilations._3, dilations._4)))
   }
 
-  /** $OpDocConv2DBackpropInput
+  /** $OpDocNNConv2DBackpropInput
     *
     * @group NNOps
     * @param  inputSizes     Integer vector representing the shape of the original input, which is a 4-D tensor.
@@ -697,7 +697,7 @@ private[api] trait NN {
       Array(dilations._1, dilations._2, dilations._3, dilations._4)))
   }
 
-  /** $OpDocConv2DBackpropFilter
+  /** $OpDocNNConv2DBackpropFilter
     *
     * @group NNOps
     * @param  input          4-D tensor whose dimension order is interpreted according to the value of `dataFormat`.
@@ -739,7 +739,7 @@ private[api] trait NN {
 
   //region Pooling Ops
 
-  /** $OpDocMaxPool
+  /** $OpDocNNMaxPool
     *
     * @group NNOps
     * @param  input      4-D tensor whose dimension order is interpreted according to the value of `dataFormat`.
@@ -764,7 +764,7 @@ private[api] trait NN {
       dataFormat.name.getBytes(StandardCharsets.ISO_8859_1)))
   }
 
-  /** $OpDocMaxPoolGrad
+  /** $OpDocNNMaxPoolGrad
     *
     * @group NNOps
     * @param  originalInput  Original input tensor.
@@ -794,7 +794,7 @@ private[api] trait NN {
       padding.name.getBytes(StandardCharsets.ISO_8859_1), dataFormat.name.getBytes(StandardCharsets.ISO_8859_1)))
   }
 
-  /** $OpDocMaxPoolGradGrad
+  /** $OpDocNNMaxPoolGradGrad
     *
     * @group NNOps
     * @param  originalInput  Original input tensor.
@@ -869,7 +869,7 @@ object NN extends NN {
 
       //region Pooling Ops
 
-      /** $OpDocMaxPool
+      /** $OpDocNNMaxPool
         *
         * @param  windowSize The size of the pooling window for each dimension of the input tensor.
         * @param  stride1    Stride of the sliding window along the second dimension of `input`.
@@ -971,7 +971,7 @@ object NN extends NN {
 
       //region Convolution Ops
 
-      /** $OpDocConv2D
+      /** $OpDocNNConv2D
         *
         * @group NNOps
         * @param  filter        4-D tensor with shape `[filterHeight, filterWidth, inChannels, outChannels]`.
@@ -1132,7 +1132,7 @@ object NN extends NN {
   private[ops] def moveAxisToEnd[D <: DataType](
       input: Tensor[D],
       axis: Int,
-      rank: Tensor[INT32]
+      rank: Tensor[INT64]
   ): Tensor[D] = {
     if (axis == -1) {
       input
@@ -1141,8 +1141,8 @@ object NN extends NN {
       Basic.transpose(
         input,
         Basic.concatenate(Seq(
-          Math.range(0, axisOutput),
-          Math.range(axisOutput + 1, rank),
+          Math.range(0L, axisOutput),
+          Math.range(axisOutput + 1L, rank),
           axisOutput), 0))
     }
   }
